refactor: Change modem.py name to cellular.py
